% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{.flatten_children}
\alias{.flatten_children}
\title{Flatten children}
\usage{
.flatten_children(x)
}
\arguments{
\item{x}{A list of children.}
}
\value{
A flat list of children.
}
\description{
Recursively unpack plain lists (but not \code{ht_tag} nodes) so users
can pass \code{list(tags$li("a"), tags$li("b"))} as a single child
argument.
}
\keyword{internal}
