% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{.make_tag}
\alias{.make_tag}
\title{Create a tag function for a normal (non-void) element}
\usage{
.make_tag(tag_name)
}
\description{
Create a tag function for a normal (non-void) element
}
\keyword{internal}
