% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{.make_void_tag}
\alias{.make_void_tag}
\title{Create a tag function for a void (self-closing) element}
\usage{
.make_void_tag(tag_name)
}
\description{
Create a tag function for a void (self-closing) element
}
\keyword{internal}
