% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_list}
\alias{tag_list}
\title{Create a tag list (sibling container)}
\usage{
tag_list(...)
}
\arguments{
\item{...}{Tags, text, or other renderable objects /// Optional.}
}
\value{
A list of class \code{"hypertext.tag.list"}.
}
\description{
Useful when you want to collect sibling nodes into a single
object without wrapping them in a parent element.
}
\examples{
tl <- tag_list(tags$p("one"), tags$p("two"))
render(tl)
}
